Moved DateUtils to API package. Added comment_id and featured_community columns to initial table create.
