+++ src/main/java/com/sublinks/sublinksapi/api/lemmy/v3/admin/mapper/AdminPurgePersonMapper.java
@@ -1,9 +1,8 @@
 package com.sublinks.sublinksapi.api.lemmy.v3.admin.mapper;
 
 import com.sublinks.sublinksapi.api.lemmy.v3.admin.models.AdminPurgePerson;
-import com.sublinks.sublinksapi.api.lemmy.v3.modlog.models.ModAdd;
 import com.sublinks.sublinksapi.moderation.dto.ModerationLog;
-import com.sublinks.sublinksapi.utils.DateUtils;
+import com.sublinks.sublinksapi.api.lemmy.v3.utils.DateUtils;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.MappingConstants;
