+++ src/main/java/com/sublinks/sublinksapi/api/lemmy/v3/comment/mappers/CommentReportMapper.java
@@ -1,12 +1,11 @@
 package com.sublinks.sublinksapi.api.lemmy.v3.comment.mappers;
 
-import com.sublinks.sublinksapi.api.lemmy.v3.comment.models.Comment;
 import com.sublinks.sublinksapi.api.lemmy.v3.comment.models.CommentReport;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.MappingConstants;
 import org.springframework.core.convert.converter.Converter;
-import com.sublinks.sublinksapi.utils.DateUtils;
+import com.sublinks.sublinksapi.api.lemmy.v3.utils.DateUtils;
 import org.springframework.lang.Nullable;
 
 @Mapper(componentModel = MappingConstants.ComponentModel.SPRING)
