+++ src/main/java/com/sublinks/sublinksapi/api/lemmy/v3/community/mappers/CommunityMapper.java
@@ -1,7 +1,7 @@
 package com.sublinks.sublinksapi.api.lemmy.v3.community.mappers;
 
 import com.sublinks.sublinksapi.api.lemmy.v3.community.models.Community;
-import com.sublinks.sublinksapi.utils.DateUtils;
+import com.sublinks.sublinksapi.api.lemmy.v3.utils.DateUtils;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.MappingConstants;
