+++ src/main/java/com/sublinks/sublinksapi/api/lemmy/v3/modlog/mappers/ModRemoveCommunityMapper.java
@@ -2,7 +2,7 @@ package com.sublinks.sublinksapi.api.lemmy.v3.modlog.mappers;
 
 import com.sublinks.sublinksapi.api.lemmy.v3.modlog.models.ModRemoveCommunity;
 import com.sublinks.sublinksapi.moderation.dto.ModerationLog;
-import com.sublinks.sublinksapi.utils.DateUtils;
+import com.sublinks.sublinksapi.api.lemmy.v3.utils.DateUtils;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.MappingConstants;
