+++ src/main/java/com/sublinks/sublinksapi/api/lemmy/v3/modlog/services/ModerationLogService.java
@@ -41,10 +41,7 @@ import com.sublinks.sublinksapi.moderation.dto.ModerationLog;
 import com.sublinks.sublinksapi.moderation.events.ModerationLogCreatedPublisher;
 import com.sublinks.sublinksapi.moderation.repositories.ModerationLogRepository;
 import com.sublinks.sublinksapi.person.repositories.PersonRepository;
-import com.sublinks.sublinksapi.post.dto.PostRead;
-import com.sublinks.sublinksapi.post.repositories.PostReportRepository;
 import com.sublinks.sublinksapi.post.repositories.PostRepository;
-import com.sublinks.sublinksapi.utils.DateUtils;
 import lombok.RequiredArgsConstructor;
 import org.springframework.core.convert.ConversionService;
 import org.springframework.data.domain.Page;
