+++ src/main/java/com/sublinks/sublinksapi/api/lemmy/v3/post/mappers/PostReportMapper.java
@@ -5,7 +5,7 @@ import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.MappingConstants;
 import org.springframework.core.convert.converter.Converter;
-import com.sublinks.sublinksapi.utils.DateUtils;
+import com.sublinks.sublinksapi.api.lemmy.v3.utils.DateUtils;
 import org.springframework.lang.Nullable;
 
 @Mapper(componentModel = MappingConstants.ComponentModel.SPRING)
