+++ src/main/java/com/sublinks/sublinksapi/api/lemmy/v3/user/mappers/PersonRegistrationApplicationMapper.java
@@ -6,7 +6,7 @@ import jakarta.annotation.Nullable;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.MappingConstants;
-import com.sublinks.sublinksapi.utils.DateUtils;
+import com.sublinks.sublinksapi.api.lemmy.v3.utils.DateUtils;
 import org.springframework.core.convert.converter.Converter;
 
 @Mapper(componentModel = MappingConstants.ComponentModel.SPRING)
